converted uint32 to uint64 for larger computations.  Optimized pascal triangle data structure.  Fixed output errors.

+++ Lambda_Partition_Algorithm/input.hpp
@@ -24,7 +24,7 @@ public:
 private:
     string input_file_name;
     vector<Term> polynomial;
-    vector<uint32_t> partition;
+    vector<uint64_t> partition;
     optional<string> POLYNOMIAL = nullopt;
     optional<string> PARTITION = nullopt;
     optional<size_t> N = nullopt;
